+++ src/EnrolmentWorkflow/Request/Repository.php
@@ -293,13 +293,14 @@ final class Repository
 
 
     /**
-     * @param int $obj_ref_id
-     * @param int $step_id
-     * @param int $user_id
+     * @param int        $obj_ref_id
+     * @param int        $step_id
+     * @param int        $user_id
+     * @param array|null $required_data
      *
      * @return Request
      */
-    public function request(int $obj_ref_id, int $step_id, int $user_id) : Request
+    public function request(int $obj_ref_id, int $step_id, int $user_id,/*?*/ array $required_data = null) : Request
     {
         $request = $this->getRequest($obj_ref_id, $step_id, $user_id);
 
@@ -319,13 +320,13 @@ final class Repository
 
                 if ($request->getStep()->getWorkflowId() === $request_backup->getStep()->getWorkflowId()) {
                     if (in_array($request->getStepId(), array_keys(self::srUserEnrolment()->enrolmentWorkflow()->steps()->getStepsForAcceptRequest($request_backup, self::dic()->user()->getId())))) {
-                        return $this->request($request->getObjRefId(), $request->getStepId(), $request->getUserId());
+                        return $this->request($request->getObjRefId(), $request->getStepId(), $request->getUserId(), $required_data);
                     }
 
                     $request = $request_backup;
                 } else {
                     if (in_array($request->getStepId(), array_keys(self::srUserEnrolment()->enrolmentWorkflow()->steps()->getStepsForAcceptRequest($request, self::dic()->user()->getId())))) {
-                        return $this->request($request->getObjRefId(), $request->getStepId(), $request->getUserId());
+                        return $this->request($request->getObjRefId(), $request->getStepId(), $request->getUserId(), $required_data);
                     }
 
                     $request = $request_backup;
@@ -334,7 +335,7 @@ final class Repository
 
             $this->storeRequest($request);
 
-            self::srUserEnrolment()->requiredData()->fills()->storeFillValues($request->getRequestId());
+            self::srUserEnrolment()->requiredData()->fills()->storeFillValues($request->getRequestId(), $required_data);
 
             self::dic()->appEventHandler()->raise(IL_COMP_PLUGIN . "/" . ilSrUserEnrolmentPlugin::PLUGIN_NAME, ilSrUserEnrolmentPlugin::EVENT_AFTER_REQUEST, [
                 "request" => $request